feat: adds nav to footer and mdularises for sharing
