+++ styles/Footer.module.sass
@@ -4,3 +4,7 @@
 
     hr
         margin-bottom: 1rem
+
+.footerContent
+    display: flex
+    justify-content: space-between
