+++ styles/Nav.module.sass
@@ -1,0 +1,8 @@
+.navItem,
+.selectedNavItem
+    margin: 0 0.5rem
+    cursor: pointer
+
+// TODO: choose button styling
+.selectedNavItem
+    color: blue
