+++ styles/Navbar.module.sass
@@ -5,16 +5,6 @@
     align-items: center
     padding: 1rem 2rem
 
-.navItem,
-.selectedNavItem
-    margin: 0 0.5rem
-    cursor: pointer
-
-// TODO: choose button styling
-.selectedNavItem
-    color: blue
-
-
 .title h3,
 .title h4
     margin: 0 0 0.1rem
